[57] Support composite IDs (#157)
* [57] Support composite IDs

* support compositeId H2DbType.findAllById

* remove point table in favor of comp_id

* better repo implementations for no-ids

* cleanup

+++ magnum/src/main/scala/com/augustnagro/magnum/ClickhouseDbType.scala
@@ -16,7 +16,8 @@ object ClickhouseDbType extends DbType:
       eElemCodecs: Seq[DbCodec[?]],
       ecElemNames: Seq[String],
       ecElemNamesSql: Seq[String],
-      idIndex: Int
+      idIndices: Seq[Int],
+      idFromProduct: Seq[Any] => ID
   )(using
       eCodec: DbCodec[E],
       ecCodec: DbCodec[EC],
@@ -29,20 +30,32 @@ object ClickhouseDbType extends DbType:
       eClassTag.runtimeClass == ecClassTag.runtimeClass,
       "ClickHouse does not support generated keys, so EC must equal E"
     )
-    val idName = eElemNamesSql(idIndex)
+    val idNames = idIndices.map(eElemNamesSql)
+    val hasId = idIndices.nonEmpty
     val selectKeys = eElemNamesSql.mkString(", ")
     val ecInsertKeys = ecElemNamesSql.mkString("(", ", ", ")")
+    val idCodecs =
+      if hasId then idIndices.map(eElemCodecs).toVector
+      else Vector(idCodec)
+
+    val idWhereClause = idNames match
+      case Seq() => "1 = 0"
+      case _ =>
+        idNames
+          .zip(idCodecs)
+          .map((name, codec) => name + " = " + codec.queryRepr)
+          .mkString(" AND ")
 
     val countSql = s"SELECT count(*) FROM $tableNameSql"
     val countQuery = Frag(countSql, Vector.empty, FragWriter.empty).query[Long]
     val existsByIdSql =
-      s"SELECT 1 FROM $tableNameSql WHERE $idName = ${idCodec.queryRepr}"
+      s"SELECT 1 FROM $tableNameSql WHERE $idWhereClause"
     val findAllSql = s"SELECT $selectKeys FROM $tableNameSql"
     val findAllQuery = Frag(findAllSql, Vector.empty, FragWriter.empty).query[E]
     val findByIdSql =
-      s"SELECT $selectKeys FROM $tableNameSql WHERE $idName = ${idCodec.queryRepr}"
+      s"SELECT $selectKeys FROM $tableNameSql WHERE $idWhereClause"
     val deleteByIdSql =
-      s"DELETE FROM $tableNameSql WHERE $idName = ${idCodec.queryRepr}"
+      s"DELETE FROM $tableNameSql WHERE $idWhereClause"
     val truncateSql = s"TRUNCATE TABLE $tableNameSql"
     val truncateUpdate =
       Frag(truncateSql, Vector.empty, FragWriter.empty).update
@@ -53,58 +66,75 @@ object ClickhouseDbType extends DbType:
       idCodec.writeSingle(id, ps, pos)
       pos + idCodec.cols.length
 
+    val existsByIdImpl: (ID, DbCon) => Boolean =
+      if hasId then
+        (id, con) =>
+          Frag(existsByIdSql, IArray(id), idWriter(id))
+            .query[Int]
+            .run()(using con)
+            .nonEmpty
+      else (_, _) => false
+
+    val findByIdImpl: (ID, DbCon) => Option[E] =
+      if hasId then
+        (id, con) =>
+          Frag(findByIdSql, IArray(id), idWriter(id))
+            .query[E]
+            .run()(using con)
+            .headOption
+      else (_, _) => None
+
+    val deleteByIdImpl: (ID, DbCon) => Unit =
+      if hasId then
+        (id, con) =>
+          Frag(deleteByIdSql, IArray(id), idWriter(id)).update
+            .run()(using con)
+          ()
+      else (_, _) => ()
+
+    val deleteAllByIdImpl: (Iterable[ID], DbCon) => BatchUpdateResult =
+      if hasId then
+        (ids, con) =>
+          given DbCon = con
+          handleQuery(deleteByIdSql, ids):
+            Using(con.connection.prepareStatement(deleteByIdSql)): ps =>
+              idCodec.write(ids, ps)
+              timed(batchUpdateResult(ps.executeBatch()))
+      else (_, _) => BatchUpdateResult.Success(0)
+
     new RepoDefaults[EC, E, ID]:
       def count(using con: DbCon): Long = countQuery.run().head
 
-      def existsById(id: ID)(using DbCon): Boolean =
-        Frag(existsByIdSql, IArray(id), idWriter(id))
-          .query[Int]
-          .run()
-          .nonEmpty
+      def existsById(id: ID)(using con: DbCon): Boolean =
+        existsByIdImpl(id, con)
 
       def findAll(using DbCon): Vector[E] = findAllQuery.run()
 
       def findAll(spec: Spec[E])(using DbCon): Vector[E] =
         SpecImpl.Default.findAll(spec, tableNameSql)
 
-      def findById(id: ID)(using DbCon): Option[E] =
-        Frag(findByIdSql, IArray(id), idWriter(id))
-          .query[E]
-          .run()
-          .headOption
+      def findById(id: ID)(using con: DbCon): Option[E] =
+        findByIdImpl(id, con)
 
       def findAllById(ids: Iterable[ID])(using DbCon): Vector[E] =
         throw UnsupportedOperationException()
 
       def delete(entity: E)(using DbCon): Unit =
-        deleteById(
-          entity
-            .asInstanceOf[Product]
-            .productElement(idIndex)
-            .asInstanceOf[ID]
-        )
+        deleteById(entityToId(entity))
 
-      def deleteById(id: ID)(using DbCon): Unit =
-        Frag(deleteByIdSql, IArray(id), idWriter(id)).update
-          .run()
+      def deleteById(id: ID)(using con: DbCon): Unit =
+        deleteByIdImpl(id, con)
 
       def truncate()(using DbCon): Unit =
         truncateUpdate.run()
 
       def deleteAll(entities: Iterable[E])(using DbCon): BatchUpdateResult =
-        deleteAllById(
-          entities.map(e =>
-            e.asInstanceOf[Product].productElement(idIndex).asInstanceOf[ID]
-          )
-        )
+        deleteAllById(entities.map(entityToId))
 
       def deleteAllById(ids: Iterable[ID])(using
           con: DbCon
       ): BatchUpdateResult =
-        handleQuery(deleteByIdSql, ids):
-          Using(con.connection.prepareStatement(deleteByIdSql)): ps =>
-            idCodec.write(ids, ps)
-            timed(batchUpdateResult(ps.executeBatch()))
+        deleteAllByIdImpl(ids, con)
 
       def insert(entityCreator: EC)(using con: DbCon): Unit =
         handleQuery(insertSql, entityCreator):
@@ -143,6 +173,11 @@ object ClickhouseDbType extends DbType:
           con: DbCon
       ): BatchUpdateResult =
         throw UnsupportedOperationException()
+
+      def entityToId(entity: E): ID =
+        val product = entity.asInstanceOf[Product]
+        val idValues = idIndices.map(i => product.productElement(i))
+        idFromProduct(idValues)
 
     end new
   end buildRepoDefaults
